fixed pagination bug
