+++ src/components/store-display/StoreDisplay.jsx
@@ -1,4 +1,5 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
+import getAllItems from "../../utils/getAllItems";
 import Items from "./Items";
 import Cart from "./cart/Cart";
 import SortItems from "./SortItems";
@@ -7,32 +8,61 @@ const StoreDisplay = () => {
   const [sortingOption, setSortingOption] = useState("");
   const [checked, setChecked] = useState(false);
   const [category, setCategory] = useState("");
+  const [allItems, setAllItems] = useState([]);
+  const [currentPage, setCurrentPage] = useState(1);
+
+  const getItems = async () => {
+    const results = await getAllItems();
+    setAllItems(results);
+  };
+
+  useEffect(() => {
+    getItems();
+  }, []);
+
+  const changePage = (pageNumber) => {
+    setCurrentPage(pageNumber);
+  };
+
+  const resetPage = () => {
+    setCurrentPage(1);
+  };
 
   const handleSorting = (option) => {
+    resetPage();
     setSortingOption(option);
   };
 
   const handleChange = () => {
+    resetPage();
     setChecked(!checked);
-  }
+  };
 
   const handleCategory = (option) => {
+    resetPage();
     setCategory(option);
-  }
+  };
 
   return (
     <div>
-      <p>Store display.</p>
       <Cart />
-      <SortItems 
-        handleSorting={handleSorting} 
-        handleChange={handleChange} 
-        sortingOption={sortingOption} 
-        checked={checked} 
-        category={category} 
-        handleCategory={handleCategory} 
+      <SortItems
+        handleSorting={handleSorting}
+        handleChange={handleChange}
+        sortingOption={sortingOption}
+        checked={checked}
+        category={category}
+        handleCategory={handleCategory}
+        resetPage={resetPage}
       />
-      <Items sortingOption={sortingOption} checked={checked} category={category} />
+      <Items
+        sortingOption={sortingOption}
+        checked={checked}
+        category={category}
+        allItems={allItems}
+        currentPage={currentPage}
+        changePage={changePage}
+      />
     </div>
   );
 };
